Fix Y2038 timestamp overflow in STP
- Change sys_get_seconds() return type from uint32_t to uint64_t
- Change topology_change_tick from UINT32 to UINT64 in BRIDGE_DATA struct
- Change topology_change_time from uint32_t to uint64_t in STP_VLAN_TABLE
- Add UINT64 type definition to l2.h

These changes prevent timestamp overflow that would occur on
January 19, 2038 when 32-bit signed timestamps exceed INT32_MAX.

+++ include/stp_dbsync.h
@@ -39,7 +39,7 @@ typedef struct {
     uint8_t forward_delay;
     uint8_t hold_time;
 
-    uint32_t topology_change_time; // time of last tc event
+    uint64_t topology_change_time; // time of last tc event (Y2038-safe)
     uint32_t topology_change_count;
 
     char root_bridge_id[STP_SYNC_BRIDGE_ID_LEN];
